third_party/WebKit/LayoutTests/external/wpt/workers/constructors/SharedWorker/null-arguments-expected.txt

This is a testharness.js-based test.
FAIL null as argumentsassert_equals: second arg expected "" but got "null"
Harness: the test ran to completion.
